// apps/web/src/api/generated/orval/model/propertyMetaResponseOptionMode.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 옵션 입력 모드
 */
export type PropertyMetaResponseOptionMode = typeof PropertyMetaResponseOptionMode[keyof typeof PropertyMetaResponseOptionMode];


export const PropertyMetaResponseOptionMode = {
  FIXED: 'FIXED',
  CREATABLE: 'CREATABLE',
} as const;