// apps/web/src/api/generated/orval/model/propertyMetaResponseValueType.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

/**
 * 값 타입
 */
export type PropertyMetaResponseValueType = typeof PropertyMetaResponseValueType[keyof typeof PropertyMetaResponseValueType];


export const PropertyMetaResponseValueType = {
  STRING: 'STRING',
  INTEGER: 'INTEGER',
  FLOAT: 'FLOAT',
  BOOLEAN: 'BOOLEAN',
  OPTION: 'OPTION',
} as const;